Finished with the QD tab for now.
The QD tab currently displays the selected questioned document, a plot of the nodes in the qd, and the cluster fill counts for the qd. These three items are displayed in a navbar page within the QD tab. I originally tried displaying them on the same page, one after the other, but I thought the plots were harded to see that way and the layout looked a little messier.