Added a tab for user to select working folder
Initially I set up the app to save files to a temp directory, but if the process is stopped for some reason, the user needs to restart from the beginning. Some files take a while to create, so losing progress isn't ideal. I changed the app to ask the user to choose a directory and files are saved to it. This way, users should be able to pick up where they left off.