Turned off the progress bars
I turned off the progress bars on the Known Writers and QD tabs because each tab only calls a single function and the progress bar can give an update before and after the function but it can't give updates for the various steps in the function. A single progress update that the function is running didn't seem helpful to me. I like the progress spinners more.