Changed QD and Known tabs' dispalys to buttons and popup windows
Displaying the items on the QD tab and Known tab as dashboard boxes didn't allow the user to click on the box to see a larger view of the item. I switched from using dashboard to bsModal. The items are now displayed as buttons. When the user clicks a button the item opens in a pop-up window. I had to change the CSS to make the pop-up windows scrollable.