Changed app to use templateK40. Temp change only.
We want people to use a "full-size" template in practice.

Ideally, I would have templateK40 be one of the templates made from 50 CSAFE and 50 CVL docs, but unless we get permission to distribute a template created from CVL docs, the template needs to be made entirely from CSAFE docs. The problem with the current templateK40 is that it was only created from LND prompts and only has 24 clusters instead of 40.

## To Do

- Try creating a template from 50 LND and 50 WOZ docs to see if that creates 40 clusters
- Follow-up with Stacy to see whether we can use the template created from CVL data