Replace open_global with reactives
open_global was a reactiveValues. I replaced it with reactives sample1(), sample2(), and slr_df(). I'm still trying to troubleshoot the bug where the images and writer profiles don't update when new files are selected. Changing to reactives didn't fix the bug, but it does make the code cleaner, so I will keep them.