Fixed bug. writerProfile and singleImage now update when input documents are changed.
I didn't know that the inputs to these modules needed to be reactives. I changed the inputs to reactives and now the modules update the displays as expeted.

## Bug
Document 1 writer profile displays correctly for the first comparison after the app is launched. But if the user selects new documents and runs a second comparison, the writer profile for the first document select does not appear. The writer profile for the second document selected does appear though.