Fixed bug where the first writer profile wouldn't display on the second comparison
Both writer profiles appeared correctly the first time an open-set comparison was done. But if one or both of the documents were changed and a second comparison was run, the writer profile for first document selected wouldn't appear. I made a separate renderUI statement just for the writer profiles and that fixed the bug.