Added license to app
Including the license as an md file created problems because then the markdown package has to be added to imports, but a note is returned on R CMD check that markdown isn't called anywhere.

Including the license as an html file with includeHTML caused a warning that the html file was a complete html file. Also, the first few lines of the file don't display correctly.

As a temporary work-around, I added a new html snippet that includes a link to the license: https://www.gnu.org/licenses/