Export `plot_writer_profiles()` and created rates data object
I received an error message that `plot_writer_profiles()` couldn't be found when I did the following:

1. Closed the handwriterApp project in Rstudio
2. Ran `devtools::install_github('CSAFE-ISU/handwriterApp')
3. `library(handwriterRF)`
4. `library(handwriterApp)`
5. `handwriterApp()`
6. Compared two documents in the Scenario 1 tab
7. Clicked generate report

To try to fix the bug, I changed `plot_writer_profiles()` to an exported function. In order to export it, it needs an example so I created the rates data frame from `handwriterRF::cfr` to use in an example.