Deleted files generated by the altdoc package
The altdoc package is great if the functions run without error, but if `render_docs()` throws an error it doesn't give any hints as to the cause of the error. `traceback()` only says there is a problem rendering. Trying to troubleshoot was taking too much time, so I created a new website in a separate repo that doesn't use altdoc.